convert to typescript
